fix problems with failures at boot time with OSDs on dmcrypt
We setup our OSDs on with dmcrypt encrypted disks for security reasons.
These devices are handled via our puppet-dmcrypt module for the nodes
directly in the same puppet agent run as the puppet-ceph run. This cause
trouble on boot since ceph::osd::device expect the blkid and osd id to
be available, otherwise a second agent run is needed to startup the OSDs.

To resolve this: added facter code to parse the needed information from
the ceph.conf (should have the state from before the boot) and
'ceph osd dump' and add some extra facts. We use these facts now as
fallback in ceph::osd::device to startup directly on boot.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>